# src/main/java/soham/spring/serivcerepository/api/ProviderApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (4.3.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package soham.spring.serivcerepository.api;

import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import soham.spring.serivcerepository.model.Provider;
import soham.spring.serivcerepository.model.ProviderRequestDTO;

import javax.validation.Valid;
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2021-02-20T15:01:30.040636+01:00[Europe/Amsterdam]")

@Validated
@Api(value = "provider", description = "the provider API")
public interface ProviderApi {

    default ProviderApiDelegate getDelegate() {
        return new ProviderApiDelegate() {};
    }
    /**
     * POST /provider : adds a Provider
     * Adds a Provider to the system
     *
     * @param providerRequestDTO Provider item to add (optional)
     * @return Successfull response with Provider info (status code 200)
     *         or Request is not correct. (status code 400)
     *         or There is no endpoint found. (status code 404)
     *         or #/components/responses/409ItemExists (status code 409)
     *         or Internal Server Error. (status code 500)
     */
    @ApiOperation(value = "adds a Provider", nickname = "addProvider", notes = "Adds a Provider to the system", response = Provider.class, tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Successfull response with Provider info", response = Provider.class),
        @ApiResponse(code = 400, message = "Request is not correct."),
        @ApiResponse(code = 404, message = "There is no endpoint found."),
        @ApiResponse(code = 409, message = "#/components/responses/409ItemExists"),
        @ApiResponse(code = 500, message = "Internal Server Error.") })
    @RequestMapping(value = "/provider",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    default ResponseEntity<Provider> addProvider(@ApiParam(value = "Provider item to add"  )  @Valid @RequestBody(required = false) ProviderRequestDTO providerRequestDTO) {
        return getDelegate().addProvider(providerRequestDTO);
    }


    /**
     * GET /provider/{id} : gets a specific provider by id
     * get by provider id
     *
     * @param id  (required)
     * @return Successfull response with Provider info (status code 200)
     *         or Request is not correct. (status code 400)
     *         or There is no endpoint found. (status code 404)
     *         or Internal Server Error. (status code 500)
     */
    @ApiOperation(value = "gets a specific provider by id", nickname = "getProviderByID", notes = "get by provider id", response = Provider.class, tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Successfull response with Provider info", response = Provider.class),
        @ApiResponse(code = 400, message = "Request is not correct."),
        @ApiResponse(code = 404, message = "There is no endpoint found."),
        @ApiResponse(code = 500, message = "Internal Server Error.") })
    @RequestMapping(value = "/provider/{id}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<Provider> getProviderByID(@ApiParam(value = "",required=true) @PathVariable("id") Integer id) {
        return getDelegate().getProviderByID(id);
    }

}