# src/main/java/soham/spring/serivcerepository/api/ProvidersApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (4.3.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package soham.spring.serivcerepository.api;

import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;
import io.swagger.annotations.ApiResponse;
import io.swagger.annotations.ApiResponses;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import soham.spring.serivcerepository.model.Provider;

import java.util.List;
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2021-02-20T15:01:30.040636+01:00[Europe/Amsterdam]")

@Validated
@Api(value = "providers", description = "the providers API")
public interface ProvidersApi {

    default ProvidersApiDelegate getDelegate() {
        return new ProvidersApiDelegate() {};
    }
    /**
     * GET /providers : lists all providers
     * list all providers
     *
     * @return all providers (status code 200)
     *         or Request is not correct. (status code 400)
     *         or There is no endpoint found. (status code 404)
     *         or Internal Server Error. (status code 500)
     */
    @ApiOperation(value = "lists all providers", nickname = "getAllProviders", notes = "list all providers", response = Provider.class, responseContainer = "List", tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "all providers", response = Provider.class, responseContainer = "List"),
        @ApiResponse(code = 400, message = "Request is not correct."),
        @ApiResponse(code = 404, message = "There is no endpoint found."),
        @ApiResponse(code = 500, message = "Internal Server Error.") })
    @RequestMapping(value = "/providers",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<List<Provider>> getAllProviders() {
        return getDelegate().getAllProviders();
    }

}