# src/main/java/soham/spring/serivcerepository/api/ProviderserviceApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (4.3.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package soham.spring.serivcerepository.api;

import soham.spring.serivcerepository.model.ProviderServices;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.CookieValue;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2021-02-20T15:01:30.040636+01:00[Europe/Amsterdam]")

@Validated
@Api(value = "providerservice", description = "the providerservice API")
public interface ProviderserviceApi {

    default ProviderserviceApiDelegate getDelegate() {
        return new ProviderserviceApiDelegate() {};
    }

    /**
     * GET /providerservice/{id} : gets a specific provider by id with Services
     * get by provider id along with all services
     *
     * @param id  (required)
     * @return Successfull response with Provider and Services info (status code 200)
     *         or Request is not correct. (status code 400)
     *         or There is no endpoint found. (status code 404)
     *         or Internal Server Error. (status code 500)
     */
    @ApiOperation(value = "gets a specific provider by id with Services", nickname = "getProviderServicesByID", notes = "get by provider id along with all services", response = ProviderServices.class, tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Successfull response with Provider and Services info", response = ProviderServices.class),
        @ApiResponse(code = 400, message = "Request is not correct."),
        @ApiResponse(code = 404, message = "There is no endpoint found."),
        @ApiResponse(code = 500, message = "Internal Server Error.") })
    @RequestMapping(value = "/providerservice/{id}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<ProviderServices> getProviderServicesByID(@ApiParam(value = "",required=true) @PathVariable("id") Integer id) {
        return getDelegate().getProviderServicesByID(id);
    }

}