# src/main/java/soham/spring/serivcerepository/api/ServiceApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (4.3.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package soham.spring.serivcerepository.api;

import soham.spring.serivcerepository.model.Service;
import soham.spring.serivcerepository.model.ServiceRequestDTO;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.CookieValue;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2021-02-20T15:01:30.040636+01:00[Europe/Amsterdam]")

@Validated
@Api(value = "service", description = "the service API")
public interface ServiceApi {

    default ServiceApiDelegate getDelegate() {
        return new ServiceApiDelegate() {};
    }

    /**
     * POST /service : adds an service
     * Adds a service to the system
     *
     * @param serviceRequestDTO Service to add (optional)
     * @return Successfull response with Service info (status code 200)
     *         or Request is not correct. (status code 400)
     *         or There is no endpoint found. (status code 404)
     *         or #/components/responses/409ItemExists (status code 409)
     *         or Internal Server Error. (status code 500)
     */
    @ApiOperation(value = "adds an service", nickname = "addService", notes = "Adds a service to the system", response = Service.class, tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Successfull response with Service info", response = Service.class),
        @ApiResponse(code = 400, message = "Request is not correct."),
        @ApiResponse(code = 404, message = "There is no endpoint found."),
        @ApiResponse(code = 409, message = "#/components/responses/409ItemExists"),
        @ApiResponse(code = 500, message = "Internal Server Error.") })
    @RequestMapping(value = "/service",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    default ResponseEntity<Service> addService(@ApiParam(value = "Service to add"  )  @Valid @RequestBody(required = false) ServiceRequestDTO serviceRequestDTO) {
        return getDelegate().addService(serviceRequestDTO);
    }


    /**
     * GET /service/{id} : gets a specific service by id
     * get by service id
     *
     * @param id  (required)
     * @return Successfull response with Service info (status code 200)
     *         or Request is not correct. (status code 400)
     *         or There is no endpoint found. (status code 404)
     *         or Internal Server Error. (status code 500)
     */
    @ApiOperation(value = "gets a specific service by id", nickname = "getServiceByID", notes = "get by service id", response = Service.class, tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Successfull response with Service info", response = Service.class),
        @ApiResponse(code = 400, message = "Request is not correct."),
        @ApiResponse(code = 404, message = "There is no endpoint found."),
        @ApiResponse(code = 500, message = "Internal Server Error.") })
    @RequestMapping(value = "/service/{id}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<Service> getServiceByID(@ApiParam(value = "",required=true) @PathVariable("id") Integer id) {
        return getDelegate().getServiceByID(id);
    }

}